pyproject.toml
[tool.poetry]
name = "boostface"
version = "0.1.0"
description = ""
authors = ["Atticuszz <1831768457@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.104.0"
supabase-py-async = "*"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
numpy = "^1.26.1"
pyturbojpeg = "^1.7.2"
filterpy = "^1.4.5"
pymilvus = "^2.3.2"
sympy = "^1.12"
milvus = "2.2.12"
opencv-python = "^4.8.1.78"
onnxruntime = "^1.16.1"
onnx = "^1.15.0"
scikit-image = "^0.22.0"
line-profiler = "^4.1.1"
setuptools = "^68.2.2"
line-profiler-pycharm = "^1.1.0"
psutil = "^5.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
